requirements.txt
# 기업 분석 보고서 시스템 - 의존성 패키지
# =========== AI 파트 ================
dspy_ai==2.4.9
wikipedia==1.4.0
sentence-transformers
toml
langchain-text-splitters
trafilatura
langchain-huggingface
qdrant-client
langchain-qdrant
numpy
litellm
diskcache
psycopg2-binary
google-genai==1.59.0
rapidfuzz

# =========== DB 파트 ================# DART API
pgvector

dart-fss==0.4.15

# Database
psycopg2-binary==2.9.9

# Web/HTML Parsing
beautifulsoup4==4.14.3
lxml==6.0.2

# Environment
python-dotenv==1.1.0

# HTTP
requests==2.32.3
httpx==0.28.1

# Utilities
tqdm==4.67.1

# Embedding (HuggingFace)
transformers>=4.36.0
torch>=2.0.0
numpy>=1.24.0

html5lib
hf_xet


# Enterprise STORM Database Layer - Updated Requirements
# New dependencies for SQLAlchemy async and Alembic

# ========== Core SQLAlchemy 2.0+ (Async) ==========
sqlalchemy==2.0.28                 # Modern ORM with async support
asyncpg==0.30.0                    # PostgreSQL async driver (critical for performance)
alembic==1.13.3                    # Database migration tool
greenlet==3.1.1                    # Required for asyncpg

# ========== Data Validation & Serialization ==========
pydantic==2.12.5                    # V2 for better performance and type support
# pydantic-settings==2.5.0           # Environment variable management

# ========== Async Web Framework ==========
fastapi==0.115.4                   # Async web framework
uvicorn==0.32.0                    # ASGI server
httpx==0.28.1                      # Async HTTP client

# ========== Development & Testing ==========
pytest==8.3.2                      # Testing framework
pytest-asyncio==0.24.0             # Async test support
pytest-cov==4.1.0                  # Coverage reporting

# ========== Legacy Dependencies (Keep for Compatibility) ==========
# AI/ML
dspy_ai==2.4.9
sentence-transformers
torch>=2.0.0
transformers>=4.36.0
numpy>=1.24.0

# DART API
dart-fss==0.4.15

# Current DB (Gradual Migration)
psycopg2-binary==2.9.9

# Web/HTML
beautifulsoup4==4.14.3
lxml==6.0.2
requests==2.32.3
trafilatura

# Utilities
python-dotenv==1.1.0
tqdm==4.67.1

# Knowledge Graph
litellm
langchain-text-splitters
langchain-huggingface
qdrant-client
langchain-qdrant
diskcache
